refactor/split commands from evm types.go into their own file
